# contracts/openapi/tool-defect-api-v1.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "刀具缺陷检测系统接口",
    "version": "1.0.0",
    "description": "采集、业务查询、复核、数据集、训练、模型发布和内部推理接口的 v1 唯一契约源。"
  },
  "servers": [
    {
      "url": "https://api.example.invalid",
      "description": "地址由环境配置提供，示例域名不可用于生产"
    }
  ],
...
  "paths": {
    "/api/v1/edge/captures": {
      "post": {
        "operationId": "createCapture",
        "tags": [
          "edge"
        ],
        "security": [
          {
            "EdgeDeviceMtls": [],
            "EdgeOAuth": [
              "capture:write"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "采集事件已创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureCreateResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "x-write-example-operation": "createCapture"
      }
    },
    "/api/v1/edge/captures/{capture_id}/images/{image_id}/complete": {
      "post": {
        "operationId": "completeCaptureImage",
        "tags": [
          "edge"
        ],
        "security": [
          {
            "EdgeDeviceMtls": [],
            "EdgeOAuth": [
              "capture:write"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CaptureId"
          },
          {
            "$ref": "#/components/parameters/ImageId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectCompleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "对象已确认，重复确认相同内容返回相同结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectCompleteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "x-write-example-operation": "completeCaptureImage"
      }
    },
...
    "/api/v1/edge/captures/{capture_id}/submit": {
      "post": {
        "operationId": "submitCapture",
        "tags": [
          "edge"
        ],
        "security": [
          {
            "EdgeDeviceMtls": [],
            "EdgeOAuth": [
              "capture:write"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CaptureId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCaptureRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "检测任务已排队",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitCaptureResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "x-write-example-operation": "submitCapture"
      }
    },
    "/api/v1/edge/captures/{capture_id}": {
      "get": {
        "operationId": "getEdgeCapture",
        "tags": [
          "edge"
        ],
        "security": [
          {
            "EdgeDeviceMtls": [],
            "EdgeOAuth": [
              "capture:read"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CaptureId"
          }
        ],
        "responses": {
          "200": {
            "description": "中央状态和最终业务处置",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureStatusResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/edge/sync/captures/query": {
      "post": {
        "operationId": "queryCaptureSync",
        "tags": [
          "edge"
        ],
        "security": [
          {
            "EdgeDeviceMtls": [],
            "EdgeOAuth": [
              "capture:read"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureSyncQueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "批量中央投影",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureSyncQueryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "x-write-example-operation": "queryCaptureSync"
      }
    },
    "/api/v1/edge/devices/{device_id}/heartbeat": {
      "post": {
        "operationId": "reportDeviceHeartbeat",
        "tags": [
          "edge"
        ],
        "security": [
          {
            "EdgeDeviceMtls": [],
            "EdgeOAuth": [
              "device:heartbeat"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DeviceId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "心跳已接收",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Acknowledgement"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "x-write-example-operation": "reportDeviceHeartbeat"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "EdgeDeviceMtls": {
        "type": "mutualTLS"
      },
      "EdgeOAuth": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://identity.example.invalid/oauth2/token",
            "scopes": {
              "capture:write": "写入采集事件",
              "capture:read": "读取采集状态",
              "device:heartbeat": "上报设备心跳"
            }
          }
        }
      },
      "UserOidc": {
        "type": "openIdConnect",
        "openIdConnectUrl": "https://identity.example.invalid/.well-known/openid-configuration"
      },
      "ServiceMtls": {
        "type": "mutualTLS"
      },
      "ServiceOAuth": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://identity.example.invalid/oauth2/token",
            "scopes": {
              "inference:callback": "提交推理执行结果",
              "runtime:read": "读取推理运行时健康"
            }
          }
        }
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 256
        }
      },
...
      "RequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "schema": {
          "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
        }
      },
      "Traceparent": {
        "name": "traceparent",
        "in": "header",
        "required": false,
        "schema": {
          "$ref": "../json-schema/common-v1.schema.json#/$defs/Traceparent"
        }
      },
      "CaptureId": {
        "name": "capture_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
        }
      },
      "ImageId": {
        "name": "image_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
        }
      },
      "DeviceId": {
        "name": "device_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
        }
      },
...
    },
...
    "responses": {
      "BadRequest": {
        "description": "请求格式错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "身份认证失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
      "Forbidden": {
        "description": "没有资源或操作权限",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在或调用者不可见",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
      "Conflict": {
        "description": "幂等摘要或乐观锁冲突",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
      "ValidationError": {
        "description": "语义或状态校验失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
...
    },
    "schemas": {
      "CaptureCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "capture_id",
          "station_id",
          "trigger",
          "recipe_id",
          "quality",
          "images"
        ],
        "properties": {
          "capture_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "station_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "trigger": {
            "$ref": "#/components/schemas/Trigger"
          },
          "recipe_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "quality": {
            "$ref": "#/components/schemas/QualitySummary"
          },
          "images": {
            "type": "array",
            "minItems": 1,
            "maxItems": 16,
            "items": {
              "$ref": "#/components/schemas/ClientImage"
            }
          }
        }
      },
      "Trigger": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "trigger_id",
          "client_sequence",
          "occurred_at",
          "source"
        ],
        "properties": {
          "trigger_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "client_sequence": {
            "type": "integer",
            "minimum": 0
          },
          "occurred_at": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/UtcTimestamp"
          },
          "source": {
            "type": "string",
            "enum": [
              "PLC",
              "SENSOR",
              "MANUAL",
              "HISTORICAL_IMPORT"
            ]
          }
        }
      },
      "QualitySummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status",
          "warnings"
        ],
        "properties": {
          "status": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/PreprocessQualityStatus"
          },
          "warnings": {
            "type": "array",
            "maxItems": 32,
            "items": {
              "type": "string",
              "maxLength": 128
            }
          }
        }
      },
      "ClientImage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "client_image_id",
          "image_role",
          "file_name",
          "media_type",
          "size_bytes",
          "sha256",
          "width",
          "height"
        ],
        "properties": {
          "client_image_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "image_role": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "file_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "pattern": "^[^/\\\\]+$"
          },
          "media_type": {
            "type": "string",
            "enum": [
              "image/png",
              "image/jpeg"
            ]
          },
          "size_bytes": {
            "type": "integer",
            "minimum": 1
          },
          "sha256": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Sha256"
          },
          "width": {
            "type": "integer",
            "minimum": 1,
            "maximum": 32768
          },
          "height": {
            "type": "integer",
            "minimum": 1,
            "maximum": 32768
          }
        }
      },
      "CaptureCreateResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "capture_id",
          "status",
          "images"
        ],
        "properties": {
          "capture_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "status": {
            "const": "UPLOADING"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadTicketResponse"
            }
          }
        }
      },
      "ObjectCompleteRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "size_bytes",
          "sha256",
          "upload_receipt"
        ],
        "properties": {
          "size_bytes": {
            "type": "integer",
            "minimum": 1
          },
          "sha256": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Sha256"
          },
          "upload_receipt": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          }
        }
      },
      "ObjectCompleteResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "image_id",
          "state",
          "sha256"
        ],
        "properties": {
          "image_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "state": {
            "const": "AVAILABLE"
          },
          "sha256": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Sha256"
          }
        }
      },
...
      "SubmitCaptureRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "requested_at"
        ],
        "properties": {
          "requested_at": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/UtcTimestamp"
          }
        }
      },
      "SubmitCaptureResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "capture_id",
          "status",
          "detection_task_id",
          "pipeline_version",
          "poll_after_ms"
        ],
        "properties": {
          "capture_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "status": {
            "const": "SUBMITTED"
          },
          "detection_task_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "pipeline_version": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Version"
          },
          "poll_after_ms": {
            "type": "integer",
            "minimum": 100,
            "maximum": 60000
          }
        }
      },
      "CaptureStatusResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "capture_id",
          "capture_status",
          "business_disposition",
          "poll_after_ms"
        ],
        "properties": {
          "capture_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "capture_status": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/CaptureStatus"
          },
          "business_disposition": {
            "oneOf": [
              {
                "$ref": "../json-schema/common-v1.schema.json#/$defs/BusinessDisposition"
              },
              {
                "type": "null"
              }
            ]
          },
          "detection": {
            "$ref": "#/components/schemas/DetectionSummary"
          },
          "review": {
            "$ref": "#/components/schemas/ReviewSummary"
          },
          "poll_after_ms": {
            "type": "integer",
            "minimum": 0,
            "maximum": 900000
          }
        }
      },
      "CaptureSyncQueryRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "capture_ids"
        ],
        "properties": {
          "capture_ids": {
            "type": "array",
            "minItems": 1,
            "maxItems": 200,
            "uniqueItems": true,
            "items": {
              "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
            }
          }
        }
      },
      "CaptureSyncQueryResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CaptureStatusResponse"
            }
          }
        }
      },
      "HeartbeatRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "agent_version",
          "reported_at",
          "queue_depth",
          "oldest_task_age_seconds",
          "disk_usage_ratio",
          "camera_status",
          "plc_status",
          "clock_offset_ms"
        ],
        "properties": {
          "agent_version": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Version"
          },
          "reported_at": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/UtcTimestamp"
          },
          "queue_depth": {
            "type": "integer",
            "minimum": 0
          },
          "oldest_task_age_seconds": {
            "type": "number",
            "minimum": 0
          },
          "disk_usage_ratio": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "camera_status": {
            "type": "string",
            "enum": [
              "ONLINE",
              "OFFLINE",
              "DEGRADED"
            ]
          },
          "plc_status": {
            "type": "string",
            "enum": [
              "ONLINE",
              "OFFLINE",
              "DEGRADED"
            ]
          },
          "clock_offset_ms": {
            "type": "number"
          }
        }
      },
...
      "UploadTicketResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "image_id",
          "upload"
        ],
        "properties": {
          "image_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "upload": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "method",
              "url",
              "headers",
              "expires_at"
            ],
            "properties": {
              "method": {
                "const": "PUT"
              },
              "url": {
                "type": "string",
                "format": "uri",
                "pattern": "^https://"
              },
              "headers": {
                "type": "object",
                "maxProperties": 8
              },
              "expires_at": {
                "$ref": "../json-schema/common-v1.schema.json#/$defs/UtcTimestamp"
              }
            }
          }
        }
      },
...
      "Acknowledgement": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "accepted",
          "request_id"
        ],
        "properties": {
          "accepted": {
            "type": "boolean"
          },
          "request_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          }
        }
      },
...
    }
  },
...
}